Move App to components, create Input component
