--- src/App.css
@@ -1,69 +1,0 @@
-/* PAGE */
-
-body,
-.pure-g [class*=pure-u] {
-  font-family: 'Open Sans', sans-serif;
-}
-
-h1, h2, h3 {
-  font-family: 'Oswald', 'Arial Narrow', sans-serif;
-}
-
-code {
-  display: block;
-  margin: 20px 0 15px 0;
-  padding: 10px;
-  background: #eee;
-}
-
-.container {
-  box-sizing: border-box;
-  width: 100%;
-  padding: 45px 20px;
-}
-
-.pure-button-primary {
-  background-color: #7f1734;
-}
-
-.pure-button-primary:hover {
-  background-color: #000000;
-}
-
-.pure-form input[type="text"]:focus {
-  border-color: #7f1734;
-}
-
-/* NAVBAR */
-
-.navbar {
-  position: fixed;
-  padding: 0px 5px 5px;
-  background: #7f1734;
-  font-family: 'Oswald', 'Arial Narrow', sans-serif;
-}
-
-.navbar a {
-  color: #fff;
-}
-
-.navbar a:active,
-.navbar a:focus,
-.navbar a:hover {
-  background: #000000;
-}
-
-.navbar img {
-  margin-top: 10px;
-  height: 40px;
-  width: 50px;
-}
-
-.navbar .pure-menu-heading {
-  font-weight: bold;
-  text-transform: none;
-}
-
-.navbar .navbar-right {
-  float: right;
-}
